apps: react to deployer and strategies move
