+++ staging/src/github.com/openshift/oc/pkg/cli/deployer/deployer_test.go
@@ -13,8 +13,8 @@ import (
 	appsv1 "github.com/openshift/api/apps/v1"
 
 	"github.com/openshift/library-go/pkg/apps/appsutil"
-	"github.com/openshift/origin/pkg/apps/strategy"
-	"github.com/openshift/origin/pkg/apps/strategy/util/appstest"
+	"github.com/openshift/oc/pkg/cli/deployer/strategy"
+	"github.com/openshift/oc/pkg/cli/deployer/strategy/util/appstest"
 )
 
 func TestDeployer_getDeploymentFail(t *testing.T) {
