+++ staging/src/github.com/openshift/oc/pkg/cli/version/version.go
