+++ test/integration/imagestream_test.go
@@ -15,7 +15,8 @@ import (
 	appsv1 "github.com/openshift/api/apps/v1"
 	imageclientv1 "github.com/openshift/client-go/image/clientset/versioned"
 	"github.com/openshift/library-go/pkg/image/imageutil"
-	stratsupport "github.com/openshift/origin/pkg/apps/strategy/support"
+	stratsupport "github.com/openshift/oc/pkg/cli/deployer/strategy/support"
+
 	imageapi "github.com/openshift/origin/pkg/image/apis/image"
 	imageclient "github.com/openshift/origin/pkg/image/generated/internalclientset"
 	imagetest "github.com/openshift/origin/pkg/image/util/testutil"
